[primitives/node] divided api extensions into submodules
