+++ primitives/node/src/api_ext/chain.rs
@@ -1,0 +1,30 @@
+use sp_core::Pair;
+use sp_runtime::MultiSignature;
+use my_node_runtime::SignedBlock;
+use primitive_types::H256;
+use substrate_api_client::Api;
+
+use crate::ApiResult;
+
+/// ApiClient extension that simplifies chain data access.
+pub trait ChainApi {
+	fn last_finalized_block(&self) -> ApiResult<Option<SignedBlock>>;
+	fn signed_block(&self, hash: Option<H256>) -> ApiResult<Option<SignedBlock>>;
+}
+
+impl<P: Pair> ChainApi for Api<P>
+	where
+		MultiSignature: From<P::Signature>
+{
+	fn last_finalized_block(&self) -> ApiResult<Option<SignedBlock>> {
+		self.get_finalized_head()?
+			.map_or_else(|| Ok(None), |hash| self.signed_block(Some(hash)))
+	}
+
+	fn signed_block(&self, hash: Option<H256>) -> ApiResult<Option<SignedBlock>> {
+		// Even though this is only a wrapper here, we want to have this in the trait
+		// to be able to be generic over the trait and mock the `signed_block` method
+		// in tests.
+		self.get_signed_block(hash)
+	}
+}
