+++ primitives/node/src/lib.rs
@@ -1,13 +1,17 @@
 #![cfg_attr(all(not(target_env = "sgx"), not(feature = "std")), no_std)]
 #![cfg_attr(target_env = "sgx", feature(rustc_private))]
 
+use codec::{Decode, Encode};
 #[cfg(feature = "sgx")]
 use sgx_tstd as std;
-
+use sp_core::H256;
 use std::vec::Vec;
 
-use codec::{Decode, Encode};
-use sp_core::H256;
+#[cfg(feature = "std")]
+mod api_ext;
+
+#[cfg(feature = "std")]
+pub use api_ext::*;
 
 pub type ShardIdentifier = H256;
 pub type BlockNumber = u32;
@@ -23,8 +27,3 @@ pub type SubstrateeConfirmCallFn = ([u8; 2], ShardIdentifier, H256, Vec<u8>);
 pub type ShieldFundsFn = ([u8; 2], Vec<u8>, u128, ShardIdentifier);
 pub type CallWorkerFn = ([u8; 2], Request);
 
-#[cfg(feature = "std")]
-pub mod api_ext;
-
-#[cfg(feature = "std")]
-pub use api_ext::*;
